refactor: BatchProcessor.ts -> batch.processor.ts
