+++ BE/src/game/service/batch.processor.ts
